[IMP] 修复jar部署

+++ src/main/java/io/choerodon/devops/infra/util/SshUtil.java
@@ -130,46 +130,52 @@ public class SshUtil {
         }
     }
 
-    public void sshStopJar(SSHClient ssh, String jarName, StringBuilder log) throws IOException {
-        StringBuilder stopJar = new StringBuilder();
-        stopJar.append(String.format("ps aux|grep %s | grep -v grep |awk '{print  $2}' |xargs kill -9 ", jarName));
-        stopJar.append(System.lineSeparator());
-        stopJar.append(String.format("cd /temp-jar && rm -f %s", jarName));
-        stopJar.append(" && cd /temp-log &&");
-        stopJar.append(String.format("rm -f %s", jarName.replace(".jar", ".log")));
-        LOGGER.info(stopJar.toString());
-        Session session = null;
-        try {
-            session = ssh.startSession();
-            final Session.Command cmd = session.exec(stopJar.toString());
-            cmd.join(WAIT_SECONDS, TimeUnit.SECONDS);
-            String logInfo = IOUtils.readFully(cmd.getInputStream()).toString();
-            String errorInfo = IOUtils.readFully(cmd.getErrorStream()).toString();
-            log.append(logInfo);
-            log.append(errorInfo);
-        } finally {
-            assert session != null;
-            session.close();
+    public void sshStopJar(SSHClient ssh, String jarName,String workingPath, StringBuilder log) throws IOException {
+        if (StringUtils.isEmpty(workingPath)) {
+            workingPath = ".";
+        } else {
+            workingPath = workingPath.endsWith("/") ? workingPath.substring(0, workingPath.length() - 1) : workingPath;
+        }
+
+        if (!StringUtils.isEmpty(jarName)) {
+            StringBuilder stopJar = new StringBuilder();
+            stopJar.append(String.format("ps aux|grep %s | grep -v grep |awk '{print  $2}' |xargs kill -9 ", jarName));
+            stopJar.append(System.lineSeparator());
+            stopJar.append(String.format("rm -f %s/temp-jar/%s", workingPath, jarName));
+            stopJar.append(String.format("rm -f %s/temp-log/%s", workingPath, jarName.replace(".jar", ".log")));
+            LOGGER.info(stopJar.toString());
+            Session session = null;
+            try {
+                session = ssh.startSession();
+                final Session.Command cmd = session.exec(stopJar.toString());
+                cmd.join(WAIT_SECONDS, TimeUnit.SECONDS);
+                String logInfo = IOUtils.readFully(cmd.getInputStream()).toString();
+                String errorInfo = IOUtils.readFully(cmd.getErrorStream()).toString();
+                log.append(logInfo);
+                log.append(errorInfo);
+            } finally {
+                assert session != null;
+                session.close();
+            }
         }
     }
 
     public void sshExec(SSHClient ssh, C7nNexusDeployDTO c7nNexusDeployDTO, String value, String workingPath, StringBuilder log) throws IOException {
         StringBuilder cmdStr = new StringBuilder();
         if (StringUtils.isEmpty(workingPath)) {
-            cmdStr.append("mkdir -p /temp/jar && ");
-            cmdStr.append("mkdir -p /temp/log && ");
+            workingPath = ".";
         } else {
             workingPath = workingPath.endsWith("/") ? workingPath.substring(0, workingPath.length() - 1) : workingPath;
-            cmdStr.append(String.format("mkdir -p %s/jar && ", workingPath));
-            cmdStr.append(String.format("mkdir -p %s/log && ", workingPath));
         }
-
+        cmdStr.append(String.format("mkdir -p %s/temp-jar && ", workingPath));
+        cmdStr.append(String.format("mkdir -p %s/temp-log && ", workingPath));
         Session session = null;
         try {
             session = ssh.startSession();
+            String jarPathAndName = workingPath + "/temp-jar/" + c7nNexusDeployDTO.getJarName();
             // 2.2
             String curlExec = String.format("curl -o %s -u %s:%s %s ",
-                    "/temp-jar/" + c7nNexusDeployDTO.getJarName(),
+                    jarPathAndName,
                     c7nNexusDeployDTO.getPullUserId(),
                     c7nNexusDeployDTO.getPullUserPassword(),
                     c7nNexusDeployDTO.getDownloadUrl());
@@ -188,7 +194,8 @@ public class SshUtil {
             }
 
             String logName = c7nNexusDeployDTO.getJarName().replace(".jar", ".log");
-            String javaJarExec = String.format("nohup %s > /temp-log/%s 2>&1 &", values.replace("${jar}", "/temp-jar/" + c7nNexusDeployDTO.getJarName()), logName);
+            String logPathAndName = workingPath + "/temp-log/" + logName;
+            String javaJarExec = String.format("nohup %s > %s 2>&1 &", values.replace("${jar}", jarPathAndName), logPathAndName);
 
             cmdStr.append(javaJarExec);
             cmdStr.append(System.lineSeparator());
